+++ playwright/Utils/helperUtils.ts
@@ -1,3 +1,10 @@
+import { Page } from '@playwright/test';
+import { getCurrentState, changeUserState } from './userStateUtils';
+import { WRAPUP_REASONS, USER_STATES, RONA_OPTIONS, LOGIN_MODE } from 'playwright/constants';
+import { submitWrapup } from './wrapupUtils';
+import { acceptExtensionCall, submitRonaPopup } from './incomingTaskUtils';
+import { loginViaAccessToken, disableMultiLogin, enableMultiLogin, initialiseWidgets, enableAllWidgets, } from './initUtils';
+import { stationLogout, telephonyLogin } from './stationLoginUtils';
 /**
  * Parses a time string in MM:SS format and converts it to total seconds
  * @param timeString - Time string in format "MM:SS" (e.g., "01:30" for 1 minute 30 seconds)
@@ -72,4 +79,439 @@ export async function waitForWebSocketReconnection(consoleMessages: string[], ti
     await new Promise((resolve) => setTimeout(resolve, 100));
   }
   return false;
+}
+
+/**
+ * Waits for a specific user state to be reached in the UI
+ * @param page - Playwright Page object
+ * @param expectedState - The expected user state to wait for
+ * @returns Promise<void>
+ * @throws Error if the expected state is not reached within the timeout
+ * @description Continuously checks the current user state until it matches the expected state or times out
+ * @example
+ * ```typescript
+ * await waitForState(page, USER_STATES.READY);
+ * ```
+ */
+
+export const waitForState = async (
+  page: Page,
+  expectedState: string,
+): Promise<void> => {
+  const start = Date.now();
+  const timeoutMs: number = 10000
+  while (true) {
+    const currentState = await getCurrentState(page);
+    if (currentState === expectedState) return;
+    if (Date.now() - start > timeoutMs) {
+      throw new Error(`Timed out waiting for state "${expectedState}", last state was "${currentState}"`);
+    }
+    await page.waitForTimeout(500); // Poll every 500ms
+  }
+};
+
+
+/**
+ * Retrieves the last state from captured logs
+ * @param capturedLogs - Array of log messages
+ * @returns Promise<string> - The last state name found in the logs, or a message if not found
+ * @description Filters logs for state change messages and extracts the last state name
+ * @example
+ * ```typescript
+ * const lastState = await getLastStateFromLogs(capturedLogs);
+ * console.log(lastState); // Outputs the last state name or a message if not found
+ * ```
+ */
+
+
+export async function getLastStateFromLogs(capturedLogs: string[]) {
+  const stateChangeLogs = capturedLogs.filter(log =>
+    log.includes('onStateChange invoked with state name:')
+  );
+
+  if (stateChangeLogs.length === 0) {
+    return 'No state change logs found';
+  }
+
+  const lastStateLog = stateChangeLogs[stateChangeLogs.length - 1];
+  const match = lastStateLog.match(/onStateChange invoked with state name:\s*(.+)$/);
+
+  if (!match) {
+    return 'No State change log found';
+  }
+
+  return match[1].trim();
+}
+
+
+/**
+ * Waits for a specific state to appear in the captured logs
+ * @param capturedLogs - Array of log messages
+ * @param expectedState - The expected state to wait for
+ * @param timeoutMs - Maximum time to wait for the state in milliseconds (default: 10000)
+ * @returns Promise<void>
+ * @throws Error if the expected state is not found within the timeout
+ * @description Continuously checks the last state in logs until it matches the expected state or times out
+ * @example
+ * ```typescript
+ * await waitForStateLogs(capturedLogs, 'READY');
+ * ```
+ */
+
+export const waitForStateLogs = async (
+  capturedLogs: string[],
+  expectedState: string,
+  timeoutMs: number = 10000
+): Promise<void> => {
+  const start = Date.now();
+  while (true) {
+    // Check if the latest state in logs matches expectedState
+    try {
+      const lastState = await getLastStateFromLogs(capturedLogs);
+      if (lastState === expectedState) return;
+    } catch {
+      // Ignore error if no state log yet
+    }
+    if (Date.now() - start > timeoutMs) {
+      throw new Error(`Timed out waiting for state "${expectedState}" in logs`);
+    }
+    await new Promise(res => setTimeout(res, 300)); // Poll every 300ms
+  }
+};
+
+
+/**
+ * Waits for a specific wrapup reason to appear in the captured logs
+ * @param capturedLogs - Array of log messages
+ * @param expectedReason - The expected wrapup reason to wait for
+ * @param timeoutMs - Maximum time to wait for the wrapup reason in milliseconds (default: 10000)
+ * @returns Promise<void>
+ * @throws Error if the expected wrapup reason is not found within the timeout
+ * @description Continuously checks the last wrapup reason in logs until it matches the expected reason or times out
+ * @example
+ * ```typescript
+ * await waitForWrapupReasonLogs(capturedLogs, 'Task Completed');
+ * ```
+ */
+
+export const waitForWrapupReasonLogs = async (
+  capturedLogs: string[],
+  expectedReason: string,
+  timeoutMs: number = 10000
+): Promise<void> => {
+  const start = Date.now();
+  while (true) {
+    try {
+      const lastReason = await getLastWrapupReasonFromLogs(capturedLogs);
+      if (lastReason === expectedReason) return;
+    } catch {
+      // Ignore error if no wrapup log yet
+    }
+    if (Date.now() - start > timeoutMs) {
+      throw new Error(`Timed out waiting for wrapup reason "${expectedReason}" in logs`);
+    }
+    await new Promise(res => setTimeout(res, 300)); // Poll every 300ms
+  }
+};
+
+
+/**
+ * Retrieves the last wrapup reason from captured logs
+ * @param capturedLogs - Array of log messages
+ * @returns Promise<string> - The last wrapup reason found in the logs, or a message if not found
+ * @description Filters logs for wrapup messages and extracts the last wrapup reason
+ * @example
+ * ```typescript
+ * const lastWrapupReason = await getLastWrapupReasonFromLogs(capturedLogs);
+ * console.log(lastWrapupReason); // Outputs the last wrapup reason or a message if not found
+ * ```
+ */
+
+
+export async function getLastWrapupReasonFromLogs(capturedLogs: string[]): Promise<string> {
+  const wrapupLogs = capturedLogs.filter(log =>
+    log.includes('onWrapup invoked with reason :')
+  );
+
+  if (wrapupLogs.length === 0) {
+    return 'No wrapup reason found';
+  }
+
+  const lastWrapupLog = wrapupLogs[wrapupLogs.length - 1];
+  const match = lastWrapupLog.match(/onWrapup invoked with reason : (.+)$/);
+
+  if (!match) {
+    return 'No wrapup reason found';
+  }
+
+  return match[1].trim();
+}
+
+/**
+ * Verifies the captured logs for wrapup and state change events
+ * @param capturedLogs - Array of log messages
+ * @param expectedWrapupReason - The expected wrapup reason to verify
+ * @param expectedState - The expected state name to verify
+ * @param shouldWrapupComeFirst - Whether the wrapup log should come before the state change log (default: true)
+ * @returns Promise<boolean> - True if verification is successful, otherwise throws an error
+ * @throws Error if logs do not match expected values or order
+ * @description Checks the last wrapup reason and state name in logs against expected values, ensuring correct order if specified
+ * @example
+ * ```typescript
+ * await verifyCallbackLogs(capturedLogs, 'Task Completed', 'READY');
+ * ```
+ */
+
+export async function verifyCallbackLogs(
+  capturedLogs: string[],
+  expectedWrapupReason: string,
+  expectedState: string,
+  shouldWrapupComeFirst: boolean = true
+): Promise<boolean> {
+  const wrapupLogs = capturedLogs.filter(log =>
+    log.includes('onWrapup invoked with reason :')
+  );
+  const stateChangeLogs = capturedLogs.filter(log =>
+    log.includes('onStateChange invoked with state name:')
+  );
+
+  if (wrapupLogs.length === 0 || stateChangeLogs.length === 0) {
+    throw new Error('Missing required logs, check callbacks for wrapup or statechange');
+  }
+
+  const lastWrapupLog = wrapupLogs[wrapupLogs.length - 1];
+  const lastStateChangeLog = stateChangeLogs[stateChangeLogs.length - 1];
+
+  const wrapupLogIndex = capturedLogs.lastIndexOf(lastWrapupLog);
+  const stateChangeLogIndex = capturedLogs.lastIndexOf(lastStateChangeLog);
+
+  if (shouldWrapupComeFirst && wrapupLogIndex >= stateChangeLogIndex) {
+    throw new Error('Wrapup log should come before state change log');
+  }
+
+  const wrapupMatch = lastWrapupLog.match(/onWrapup invoked with reason : (.+)$/);
+  const stateMatch = lastStateChangeLog.match(/onStateChange invoked with state name:\s*(.+)$/);
+
+  if (!wrapupMatch || !stateMatch) {
+    throw new Error('Could not extract values from logs');
+  }
+
+  const actualWrapupReason = wrapupMatch[1].trim();
+  const actualStateName = stateMatch[1].trim();
+
+  // Verify expected values
+  if (actualWrapupReason !== expectedWrapupReason) {
+    throw new Error('Wrapup reason mismatch, expected ' + expectedWrapupReason + ', got ' + actualWrapupReason);
+  }
+
+  if (actualStateName !== expectedState) {
+    throw new Error('State name mismatch, expected ' + expectedState + ', got ' + actualStateName);
+  }
+
+  return true;
+}
+
+/**
+ * Compares two RGB color strings to check if they are within a specified tolerance
+ * @param receivedColor - The color received from the UI (e.g., "rgb(255, 0, 0)")
+ * @param expectedColor - The expected color to compare against (e.g., "rgb(250, 5, 0)")
+ * @param tolerance - The maximum allowed difference for each RGB component (default: 5)
+ * @returns boolean - True if colors are close enough, false otherwise
+ * @description Compares each RGB component of the two colors and checks if the absolute difference is within the specified tolerance
+ * @example
+ * ```typescript
+ * const isClose = isColorClose("rgb(255, 0, 0)", "rgb(250, 5, 0)");
+ * console.log(isClose); // Outputs true or false based on comparison
+ * ```
+ */
+
+export function isColorClose(
+  receivedColor: string,
+  expectedColor: string,
+  tolerance: number = 5
+): boolean {
+  console.log(`Comparing colors: received=${receivedColor}, expected=${expectedColor}, tolerance=${tolerance}`);
+  const receivedRgb = receivedColor.match(/\d+/g)?.map(Number) || [];
+  const expectedRgb = expectedColor.match(/\d+/g)?.map(Number) || [];
+
+  for (let i = 0; i < 3; i++) {
+    if (
+      typeof receivedRgb[i] !== 'number' ||
+      typeof expectedRgb[i] !== 'number'
+    ) {
+      continue; // skip if not present
+    }
+    if (Math.abs(receivedRgb[i] - expectedRgb[i]) > tolerance) {
+      return false;
+    }
+  }
+  return true;
+}
+
+
+/**
+ * Handles stray incoming tasks by accepting them and performing wrap-up actions, to be used for clean up before tests
+ * @param page - Playwright Page object
+ * @param extensionPage - Optional extension page for handling calls (default: null)
+ * @returns Promise<void>
+ * @description Continuously checks for incoming tasks, accepts them, and performs wrap-up actions until no more tasks are available
+ * @example
+ * ```typescript
+ * await handleStrayTasks(page, extensionPage);
+ * ```
+ */
+
+export const handleStrayTasks = async (page: Page, extensionPage: Page | null = null): Promise<void> => {
+  await page.waitForTimeout(1000);
+  const incomingTaskDiv = page.getByTestId(/^samples:incoming-task(-\w+)?$/);
+
+  while (true) {
+    let flag1 = false;
+    let flag2 = true;
+    while (true) {
+      const task = incomingTaskDiv.first();
+      let isTaskVisible = await task.isVisible().catch(() => false);
+      if (!isTaskVisible) break;
+      const acceptButton = task.getByTestId('task:accept-button').first();
+      const acceptButtonVisible = await acceptButton.isVisible().catch(() => false);
+      const isExtensionCall = await (await task.innerText()).includes('Ringing...');
+      if (isExtensionCall) {
+        const extensionCallVisible = await extensionPage.locator('[data-test="right-action-button"]').waitFor({ state: 'visible', timeout: 40000 }).then(() => true).catch(() => false);
+        if (extensionCallVisible) {
+          await acceptExtensionCall(extensionPage);
+          flag1 = true
+        } else {
+          throw new Error('Accept button not visible and extension page is not available');
+        }
+      } else {
+        try {
+          await acceptButton.click({ timeout: 5000 });
+        } catch (error) { }
+        flag1 = true;
+      }
+      await page.waitForTimeout(1000);
+    }
+    const endButton = page.getByTestId(/^call-control:end-\w+$/).first();
+    const endButtonVisible = await endButton.waitFor({ state: 'visible', timeout: 2000 }).then(() => true).catch(() => false);
+    if (endButtonVisible) {
+      await page.waitForTimeout(2000);
+      await endButton.click({ timeout: 5000 });
+      await submitWrapup(page, WRAPUP_REASONS.SALE);
+    } else {
+      const wrapupBox = page.getByTestId('wrapup-button').first();
+      const isWrapupBoxVisible = await wrapupBox.waitFor({ state: 'visible', timeout: 2000 }).then(() => true).catch(() => false);
+      if (isWrapupBoxVisible) {
+        await page.waitForTimeout(2000);
+        await submitWrapup(page, WRAPUP_REASONS.SALE);
+        await page.waitForTimeout(2000)
+      } else {
+        flag2 = false;
+      }
+    }
+
+    if (!flag1 && !flag2) {
+      break;
+    }
+  }
+
+}
+
+/*
+/ * Sets up the page for testing by logging in, enabling widgets, and handling user states, cleaning up stray tasks, submitting RONA popups
+ * @param page - Playwright Page object
+ * @param loginMode - The login mode to use (e.g., LOGIN_MODE.DESKTOP or LOGIN_MODE.EXTENSION)
+ * @param extensionPage - Optional extension page for handling calls in extension mode (default: null)
+ * @returns Promise<void>
+ * @description Logs in via access token, enables all widgets, handles multi-login settings, initializes widgets, and manages user states
+ * @example
+ * ```typescript
+ * await pageSetup(page, LOGIN_MODE.DESKTOP);
+ * ```
+ */
+
+export const pageSetup = async (page: Page, loginMode: string, extensionPage: Page | null = null) => {
+  const maxRetries = 3;
+  await loginViaAccessToken(page, 'AGENT1');
+  await enableAllWidgets(page);
+  if (loginMode === LOGIN_MODE.DESKTOP) {
+    await disableMultiLogin(page);
+  } else {
+    await enableMultiLogin(page);
+  }
+
+  for (let i = 0; i < maxRetries; i++) {
+    try {
+      await initialiseWidgets(page);
+      break;
+    } catch (error) {
+      if (i == maxRetries - 1) {
+        throw new Error(`Failed to initialise widgets after ${maxRetries} attempts: ${error}`);
+      }
+    }
+  }
+
+  let loginButtonExists = await page
+    .getByTestId('login-button')
+    .isVisible()
+    .catch(() => false);
+  if (loginButtonExists) {
+    await telephonyLogin(page, loginMode);
+  } else {
+    const stateSelectVisible = await page.getByTestId('state-select').waitFor({ state: 'visible', timeout: 30000 }).then(() => true).catch(() => false);
+    if (stateSelectVisible) {
+      const ronapopupVisible = await page.getByTestId('samples:rona-popup').waitFor({ state: 'visible', timeout: 5000 }).then(() => true).catch(() => false);
+      if (ronapopupVisible) {
+        await submitRonaPopup(page, RONA_OPTIONS.AVAILABLE);
+      }
+      const userState = await getCurrentState(page);
+      await changeUserState(page, USER_STATES.AVAILABLE);
+      await page.waitForTimeout(5000);
+
+      const incomingTaskDiv = page.getByTestId(/^samples:incoming-task(-\w+)?$/).first();
+      await incomingTaskDiv.waitFor({ state: 'visible', timeout: 5000 }).catch(() => false);
+      await handleStrayTasks(page, extensionPage);
+    }
+    loginButtonExists = await page
+      .getByTestId('login-button')
+      .isVisible()
+      .catch(() => false);
+
+
+    if (!loginButtonExists) await stationLogout(page);
+    await telephonyLogin(page, loginMode);
+  }
+
+  let ronapopupVisible = await page.getByTestId('samples:rona-popup').waitFor({ state: 'visible', timeout: 5000 }).then(() => true).catch(() => false);
+  if (ronapopupVisible) {
+    await submitRonaPopup(page, RONA_OPTIONS.AVAILABLE);
+  }
+
+  let stationLoginFailure = await page.getByTestId('station-login-failure-label').waitFor({ state: 'visible', timeout: 5000 }).then(() => true).catch(() => false);
+  for (let i = 0; i < maxRetries && stationLoginFailure; i++) {
+    loginButtonExists = await page
+      .getByTestId('login-button')
+      .isVisible()
+      .catch(() => false);
+    if (!loginButtonExists) await stationLogout(page);
+    await telephonyLogin(page, loginMode);
+    await page.getByTestId('state-select').waitFor({ state: 'visible', timeout: 30000 });
+    stationLoginFailure = await page.getByTestId('station-login-failure-label').waitFor({ state: 'visible', timeout: 5000 }).then(() => true).catch(() => false);
+    if (i == maxRetries - 1 && stationLoginFailure) {
+      throw new Error(`Station Login Error Persists after ${maxRetries} attempts`);
+    }
+  }
+
+  await page.getByTestId('state-select').waitFor({ state: 'visible', timeout: 30000 });
+
+  ronapopupVisible = await page.getByTestId('samples:rona-popup').waitFor({ state: 'visible', timeout: 3000 }).then(() => true).catch(() => false);
+  if (ronapopupVisible) {
+    await submitRonaPopup(page, RONA_OPTIONS.AVAILABLE);
+  }
+
+  await changeUserState(page, USER_STATES.AVAILABLE);
+  await page.waitForTimeout(3000);
+  const incomingTaskDiv = page.getByTestId(/^samples:incoming-task(-\w+)?$/).first();
+  await incomingTaskDiv.waitFor({ state: 'visible', timeout: 5000 }).catch(() => false);
+  await handleStrayTasks(page, extensionPage);
+
 }
